refactor(lsp): extract formatting handlers into language/formatting
Move handle_formatting, handle_range_formatting, and handle_on_type_formatting
to the new language/formatting.rs module.

Also removes unused handle_on_type_formatting_old legacy delegator.
