+++ crates/perl-parser/src/lsp/server_impl/language/mod.rs
@@ -6,8 +6,10 @@
 //! - navigation: Go-to-definition, declaration, type definition, implementation
 //! - references: Find references and document highlights
 //! - symbols: Document symbols and folding ranges
+//! - formatting: Document and range formatting
 
 mod completion;
+mod formatting;
 mod hover;
 mod navigation;
 mod references;
