+++ crates/perl-parser/src/lsp/server_impl/language/symbols.rs
@@ -286,23 +286,20 @@ impl LspServer {
 /// Map symbol kind to LSP SymbolKind numeric value
 fn symbol_kind_to_lsp(kind: crate::symbol::SymbolKind) -> u32 {
     match kind {
-        crate::symbol::SymbolKind::Package => 4,          // Module
-        crate::symbol::SymbolKind::Subroutine => 12,      // Function
-        crate::symbol::SymbolKind::ScalarVariable => 13,  // Variable
-        crate::symbol::SymbolKind::ArrayVariable => 18,   // Array
-        crate::symbol::SymbolKind::HashVariable => 19,    // Object (closest match)
-        crate::symbol::SymbolKind::Constant => 14,        // Constant
-        crate::symbol::SymbolKind::Label => 16,           // String (closest match)
-        crate::symbol::SymbolKind::Format => 12,          // Function
+        crate::symbol::SymbolKind::Package => 4,         // Module
+        crate::symbol::SymbolKind::Subroutine => 12,     // Function
+        crate::symbol::SymbolKind::ScalarVariable => 13, // Variable
+        crate::symbol::SymbolKind::ArrayVariable => 18,  // Array
+        crate::symbol::SymbolKind::HashVariable => 19,   // Object (closest match)
+        crate::symbol::SymbolKind::Constant => 14,       // Constant
+        crate::symbol::SymbolKind::Label => 16,          // String (closest match)
+        crate::symbol::SymbolKind::Format => 12,         // Function
     }
 }
 
 /// Helper function to convert offset to line number
 fn offset_to_line(content: &str, offset: usize) -> usize {
-    content[..offset.min(content.len())]
-        .chars()
-        .filter(|&c| c == '\n')
-        .count()
+    content[..offset.min(content.len())].chars().filter(|&c| c == '\n').count()
 }
 
 /// Fallback folding extraction using text-based analysis
