+++ crates/perl-parser/src/lsp/server_impl/mod.rs
@@ -1294,45 +1294,12 @@ impl LspServer {
         }
     }
 
-    /// Handle textDocument/onTypeFormatting request
-    fn handle_on_type_formatting(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(p) = params {
-            let uri = p["textDocument"]["uri"].as_str().ok_or_else(|| JsonRpcError {
-                code: INVALID_PARAMS,
-                message: "Missing textDocument.uri".into(),
-                data: None,
-            })?;
-            let ch = p["ch"].as_str().and_then(|s| s.chars().next()).unwrap_or('\n');
-            let pos = &p["position"];
-            let line = pos["line"].as_u64().unwrap_or(0) as u32;
-            let col = pos["character"].as_u64().unwrap_or(0) as u32;
-
-            let documents = self.documents.lock().unwrap();
-            let doc = self.get_document(&documents, uri).ok_or_else(|| JsonRpcError {
-                code: INVALID_REQUEST,
-                message: format!("Document not open: {}", uri),
-                data: None,
-            })?;
-
-            if let Some(edits) =
-                crate::on_type_formatting::compute_on_type_edit(&doc.text, line, col, ch)
-            {
-                return Ok(Some(json!(edits)));
-            }
-        }
-        Ok(Some(json!([])))
-    }
-
     /// Get workspace roots from initialization
     fn workspace_roots(&self) -> Vec<url::Url> {
         // In a real implementation, store these from initialize params
         // For now, return empty vec
         vec![]
     }
-
 
     /// Validate if a string is a valid Perl identifier
     fn is_valid_identifier(&self, name: &str) -> bool {
@@ -1598,139 +1565,6 @@ impl LspServer {
     #[inline]
     fn offset_to_pos16(&self, doc: &DocumentState, offset: usize) -> (u32, u32) {
         doc.line_starts.offset_to_position_rope(&doc.rope, offset)
-    }
-
-    /// Handle textDocument/formatting request
-    fn handle_formatting(&self, params: Option<Value>) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-
-            // Reject stale requests
-            let req_version = params["textDocument"]["version"].as_i64().map(|n| n as i32);
-            self.ensure_latest(uri, req_version)?;
-
-            let options: FormattingOptions = serde_json::from_value(params["options"].clone())
-                .unwrap_or(FormattingOptions {
-                    tab_size: 4,
-                    insert_spaces: true,
-                    trim_trailing_whitespace: None,
-                    insert_final_newline: None,
-                    trim_final_newlines: None,
-                });
-
-            eprintln!("Formatting document: {}", uri);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let formatter = CodeFormatter::new();
-                match formatter.format_document(&doc.text, &options) {
-                    Ok(edits) => {
-                        let doc_end = self.get_document_end_position(&doc.text);
-                        let lsp_edits: Vec<Value> = edits
-                            .into_iter()
-                            .map(|edit| {
-                                json!({
-                                    "range": {
-                                        "start": {
-                                            "line": edit.range.start.line,
-                                            "character": edit.range.start.character,
-                                        },
-                                        "end": doc_end.clone(),
-                                    },
-                                    "newText": edit.new_text,
-                                })
-                            })
-                            .collect();
-
-                        return Ok(Some(json!(lsp_edits)));
-                    }
-                    Err(e) => {
-                        eprintln!("Formatting error: {}", e);
-                        return Err(JsonRpcError {
-                            code: -32603,
-                            message: format!("Formatting failed: {}", e),
-                            data: None,
-                        });
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
-    }
-
-    /// Handle textDocument/rangeFormatting request
-    fn handle_range_formatting(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        if let Some(params) = params {
-            let uri = params["textDocument"]["uri"].as_str().unwrap_or("");
-            let options: FormattingOptions = serde_json::from_value(params["options"].clone())
-                .unwrap_or(FormattingOptions {
-                    tab_size: 4,
-                    insert_spaces: true,
-                    trim_trailing_whitespace: None,
-                    insert_final_newline: None,
-                    trim_final_newlines: None,
-                });
-
-            let range = if let Some(range_value) = params.get("range") {
-                crate::formatting::Range {
-                    start: crate::formatting::Position {
-                        line: range_value["start"]["line"].as_u64().unwrap_or(0) as u32,
-                        character: range_value["start"]["character"].as_u64().unwrap_or(0) as u32,
-                    },
-                    end: crate::formatting::Position {
-                        line: range_value["end"]["line"].as_u64().unwrap_or(0) as u32,
-                        character: range_value["end"]["character"].as_u64().unwrap_or(0) as u32,
-                    },
-                }
-            } else {
-                return Ok(Some(json!([])));
-            };
-
-            eprintln!("Formatting range in document: {}", uri);
-
-            let documents = self.documents.lock().unwrap();
-            if let Some(doc) = self.get_document(&documents, uri) {
-                let formatter = CodeFormatter::new();
-                match formatter.format_range(&doc.text, &range, &options) {
-                    Ok(edits) => {
-                        let lsp_edits: Vec<Value> = edits
-                            .into_iter()
-                            .map(|edit| {
-                                json!({
-                                    "range": {
-                                        "start": {
-                                            "line": edit.range.start.line,
-                                            "character": edit.range.start.character,
-                                        },
-                                        "end": {
-                                            "line": edit.range.end.line,
-                                            "character": edit.range.end.character,
-                                        },
-                                    },
-                                    "newText": edit.new_text,
-                                })
-                            })
-                            .collect();
-
-                        return Ok(Some(json!(lsp_edits)));
-                    }
-                    Err(e) => {
-                        eprintln!("Range formatting error: {}", e);
-                        return Err(JsonRpcError {
-                            code: -32603,
-                            message: format!("Range formatting failed: {}", e),
-                            data: None,
-                        });
-                    }
-                }
-            }
-        }
-
-        Ok(Some(json!([])))
     }
 
     /// Handle textDocument/codeLens request
@@ -3758,15 +3592,6 @@ impl LspServer {
         } else {
             Ok(Some(Value::Null))
         }
-    }
-
-    /// Legacy onTypeFormatting handler retained for compatibility.
-    /// Delegates to the modern formatting pipeline.
-    fn handle_on_type_formatting_old(
-        &self,
-        params: Option<Value>,
-    ) -> Result<Option<Value>, JsonRpcError> {
-        self.handle_on_type_formatting(params)
     }
 }
